requirements.txt
Flask==3.0.0
pipreqs==0.4.13
protobuf==4.25.1
gunicorn==21.2.0
google-cloud-storage==2.13.0
google-cloud-bigquery==3.14.0
PyPDF2==3.0.1
requests==2.31.0
pandas==2.1.0
pandas-gbq==0.19.2